# spec/api/rest/OpenAPI3.yaml
openapi: 3.0.3
info:
  title: Auth Haven REST API
  description: REST API for browser UI clients. Handles authentication, registration, and user management.
  version: 1.0.0
servers:
  - url: https://api.auth-haven.com/v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: ErrInvalidCredentials
        message:
          type: string
          example: Invalid email or password

    # -------------------------
    # Tenant
    # -------------------------
...
    User:
      type: object
      properties:
        user_id:
          type: string
        tenant_id:
          type: string
        role_id:
          type: integer
          format: int64
          nullable: true
        email:
          type: string
        full_name:
          type: string
        status:
          type: integer
          description: 1=active, 2=pending, 3=disabled
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true

    # -------------------------
    # Auth
    # -------------------------
    LoginRequest:
      type: object
      required: [tenant_id, email, password]
      properties:
        tenant_id:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        mfa_required:
          type: boolean
        temp_token:
          type: string
          nullable: true
          description: Populated only when mfa_required is true
        access_token:
          type: string
          nullable: true
          description: Populated only when mfa_required is false
        refresh_token:
          type: string
          nullable: true
          description: Populated only when mfa_required is false
        session_id:
          type: string
          nullable: true
          description: Populated only when mfa_required is false
    VerifyMFARequest:
      type: object
      required: [temp_token, code]
      properties:
        temp_token:
          type: string
        code:
          type: string
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # -------------------------
    # Registration
    # -------------------------
    RegisterIndividualRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        full_name:
          type: string
    RegisterOrgUserRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        full_name:
          type: string
    RegisterWithInvitationRequest:
      type: object
      required: [invitation_token, password, full_name]
      properties:
        invitation_token:
          type: string
        password:
          type: string
          format: password
        full_name:
          type: string

    # -------------------------
    # Password
    # -------------------------
    RequestPasswordResetRequest:
      type: object
      required: [tenant_id, email]
      properties:
        tenant_id:
          type: string
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          format: password
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password

    # -------------------------
    # MFA
    # -------------------------
...
    ActivateTOTPRequest:
      type: object
      required: [mfa_id, code]
      properties:
        mfa_id:
          type: string
        code:
          type: string
...
    Invitation:
      type: object
      properties:
        invitation_id:
          type: string
        tenant_id:
          type: string
        role_id:
          type: integer
          format: int64
        email:
          type: string
        status:
          type: integer
          description: 1=pending, 2=accepted, 3=expired
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    SendInvitationRequest:
      type: object
      required: [email, role_id]
      properties:
        email:
          type: string
          format: email
        role_id:
          type: integer
          format: int64

    # -------------------------
    # Audit
    # -------------------------
    AuditLog:
      type: object
      properties:
        log_id:
          type: string
        user_id:
          type: string
        tenant_id:
          type: string
        action:
          type: string
        ip_address:
          type: string
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time
    PaginatedAuditLogs:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

# ==========================================
# PATHS
# ==========================================
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User or tenant disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify MFA code and complete login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyMFARequest'
      responses:
        '200':
          description: MFA verified, tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired temp token or MFA code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      responses:
        '204':
          description: Logged out successfully
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # Registration
  # -------------------------
  /auth/register/individual:
    post:
      tags: [Registration]
      summary: Register as an individual user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterIndividualRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register/org:
    post:
      tags: [Registration]
      summary: Register as an organization user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterOrgUserRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Public domain not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered in this tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register/invitation:
    post:
      tags: [Registration]
      summary: Register via invitation token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWithInvitationRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid or expired invitation token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered in this tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # Password
  # -------------------------
  /auth/password/forgot:
    post:
      tags: [Password]
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '204':
          description: Always returns success to prevent enumeration
  /auth/password/reset:
    post:
      tags: [Password]
      summary: Reset password using token from email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset successfully
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Password too weak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/password/change:
    post:
      tags: [Password]
      summary: Change password for authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: New password too weak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # MFA
  # -------------------------
...
  /me/mfa/totp/activate:
    post:
      tags: [MFA]
      summary: Activate TOTP after verifying code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateTOTPRequest'
      responses:
        '204':
          description: TOTP activated
        '400':
          description: Invalid MFA code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /me/mfa/{mfa_id}:
    delete:
      tags: [MFA]
      summary: Disable and remove an MFA method
      parameters:
        - name: mfa_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: MFA method removed
        '404':
          description: MFA method not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # Profile
  # -------------------------
  /me:
    get:
      tags: [Profile]
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Profile]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # -------------------------
  # Sessions
  # -------------------------
...
  /me/sessions/{session_id}:
    delete:
      tags: [Sessions]
      summary: Revoke a specific session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Session revoked
        '403':
          description: Session does not belong to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # Devices
  # -------------------------
...
  /me/devices/{device_id}:
    delete:
      tags: [Devices]
      summary: Remove a device
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Device removed
        '403':
          description: Device does not belong to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------------
  # Users (Admin)
  # -------------------------
...
  /tenants/{tenant_id}/invitations:
    get:
      tags: [Invitations]
      summary: List invitations for a tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Invitations]
      summary: Send an invitation (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationRequest'
      responses:
        '201':
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tenants/{tenant_id}/audit-logs:
    get:
      tags: [Audit]
      summary: List audit logs for a tenant (admin only)
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/audit-logs:
    get:
      tags: [Audit]
      summary: List audit logs for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'